networksecurity/utils/main_utils/utils.py
import yaml
import pickle
import dill 
import os, sys
import numpy as np 

from networksecurity.exception.exception import NetworkSecurityException
from networksecurity.logger.logger import logging

def read_yaml(file_path: str) -> dict:
    try:
        with open(file_path, "rb") as yaml_file:
            return yaml.safe_load(yaml_file)
    except Exception as e:
        NetworkSecurityException(e, sys)
    