requirements.txt
pymongo
pandas
numpy
matplotlib
python-dotenv
certifi
pymongo[srv]
pyyaml
dill
scikit-learn
-e .